+++ alphastats/DataSet_Statistics.py
@@ -17,7 +17,7 @@ class Statistics:
     def _calculate_foldchange(self, mat_transpose:pd.DataFrame, group1_samples:list, group2_samples:list):
         mat_transpose += 0.00001
 
-        if self.dataset.preprocessing_info["Log2-transformed"]:
+        if self.preprocessing_info["Log2-transformed"]:
             fc = (
                 mat_transpose[group1_samples].T.mean().values
                 - mat_transpose[group2_samples].T.mean().values