+++ alphastats/multicova/multicova.py
@@ -470,7 +470,7 @@ def get_fdr_line_regression(t_limits, s0, X, plot = False,
         s_df_neg = s_df.copy()
         s_df_neg.fc_s = -s_df_neg.fc_s
 
-        s_df = s_df.append(s_df_neg)
+        s_df = s_df.concat(s_df_neg)
 
         if (plot):
             fig = px.scatter(x=s_df.fc_s,